VideoGeneration/requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.4.2
av==10.0.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
controlnet-aux==0.0.7
croniter==1.4.1
cycler==0.12.1
datasets==2.14.5
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.6.0
defusedxml==0.7.1
diffusers @ git+https://github.com/huggingface/diffusers.git@e5168588864d72a4dca37e90318c6b11da0eaaf1
dill==0.3.7
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.2
fastjsonschema==2.18.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gpustat==1.1.1
h11==0.14.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.3
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.0.9.post0
lightning-cloud==0.5.39
lightning-utilities==0.9.0
lit==15.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.7
minexr==1.0.2
mistune==3.0.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
ninja==1.11.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.24.1
nvidia-ml-py==12.535.133
omegaconf==2.3.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
OpenEXR==1.3.8
optimum==1.13.2
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyEXR @ git+https://github.com/tvogels/pyexr.git@900fe6a95b185219a40fa74d204be05fe8c9fe31
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.9.post0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
readchar==4.0.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rouge==1.0.1
rpds-py==0.10.3
sacremoses==0.0.53
safetensors==0.3.3
scikit-image==0.22.0
scipy==1.11.2
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
terminado==0.17.1
tifffile==2023.9.26
tiktoken==0.5.1
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchmetrics==1.2.0
torchvision==0.15.2+cu117
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.33.2
transformers-stream-generator==0.0.4
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.23.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
widgetsnbextension==4.0.9
xformers==0.0.21
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
